# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Standup Comedy Connect API",
    "description": "API pour la plateforme de mise en relation entre humoristes et organisateurs d'événements de stand-up comedy",
    "version": "1.0.0",
    "contact": {
      "name": "Support API",
      "email": "support@standup-comedy-connect.com"
    }
  },
  "servers": [
    {
      "url": "https://standup-comedy-app.onrender.com/api",
      "description": "Serveur de production"
    },
    {
      "url": "http://localhost:5000/api",
      "description": "Serveur de développement local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Récupérer tous les événements",
        "description": "Retourne la liste des événements selon le rôle de l'utilisateur",
        "parameters": [
          {
            "name": "organizerId",
            "in": "query",
            "description": "ID de l'organisateur pour filtrer",
            "required": false,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des événements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {"type": "string"},
                      "title": {"type": "string"},
                      "description": {"type": "string"},
                      "date": {"type": "string", "format": "date"},
                      "startTime": {"type": "string"},
                      "endTime": {"type": "string"},
                      "location": {
                        "type": "object",
                        "properties": {
                          "venue": {"type": "string"},
                          "address": {"type": "string"},
                          "city": {"type": "string"},
                          "country": {"type": "string"}
                        }
                      },
                      "organizer": {
                        "type": "object",
                        "properties": {
                          "_id": {"type": "string"},
                          "firstName": {"type": "string"},
                          "lastName": {"type": "string"},
                          "email": {"type": "string"}
                        }
                      },
                      "status": {"type": "string", "enum": ["draft", "published", "cancelled", "completed"]},
                      "requirements": {
                        "type": "object",
                        "properties": {
                          "minExperience": {"type": "number"},
                          "maxPerformers": {"type": "number"},
                          "duration": {"type": "number"}
                        }
                      },
                      "budget": {
                        "type": "object",
                        "properties": {
                          "min": {"type": "number"},
                          "max": {"type": "number"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Créer un nouvel événement",
        "description": "Permet à un organisateur de créer un nouvel événement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "description", "date", "location", "requirements"],
                "properties": {
                  "title": {"type": "string", "minLength": 3, "maxLength": 100},
                  "description": {"type": "string", "minLength": 10, "maxLength": 1000},
                  "date": {"type": "string", "format": "date"},
                  "startTime": {"type": "string", "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"},
                  "endTime": {"type": "string", "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"},
                  "location": {
                    "type": "object",
                    "required": ["venue", "address", "city", "country"],
                    "properties": {
                      "venue": {"type": "string"},
                      "address": {"type": "string"},
                      "city": {"type": "string"},
                      "country": {"type": "string"}
                    }
                  },
                  "requirements": {
                    "type": "object",
                    "required": ["minExperience", "duration"],
                    "properties": {
                      "minExperience": {"type": "number", "minimum": 0},
                      "maxPerformers": {"type": "number", "minimum": 1},
                      "duration": {"type": "number", "minimum": 5}
                    }
                  },
                  "budget": {
                    "type": "object",
                    "properties": {
                      "min": {"type": "number", "minimum": 0},
                      "max": {"type": "number", "minimum": 0}
                    }
                  },
                  "status": {"type": "string", "enum": ["draft", "published"], "default": "draft"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Événement créé avec succès"
          },
          "400": {
            "description": "Données invalides"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "tags": ["Events"],
        "summary": "Récupérer un événement par ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Détails de l'événement"
          },
          "404": {
            "description": "Événement non trouvé"
          }
        }
      },
      "put": {
        "tags": ["Events"],
        "summary": "Modifier un événement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "description": {"type": "string"},
                  "date": {"type": "string", "format": "date"},
                  "startTime": {"type": "string"},
                  "endTime": {"type": "string"},
                  "location": {"type": "object"},
                  "requirements": {"type": "object"},
                  "budget": {"type": "object"},
                  "status": {"type": "string", "enum": ["draft", "published", "cancelled", "completed", "DRAFT", "PUBLISHED", "CANCELLED", "COMPLETED"]},
                  "cancellationReason": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Événement modifié avec succès"
          },
          "403": {
            "description": "Non autorisé à modifier cet événement"
          },
          "404": {
            "description": "Événement non trouvé"
          }
        }
      },
      "delete": {
        "tags": ["Events"],
        "summary": "Supprimer un événement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Événement supprimé avec succès"
          },
          "403": {
            "description": "Non autorisé à supprimer cet événement"
          },
          "404": {
            "description": "Événement non trouvé"
          }
        }
      }
    },
...
    "/absences/event/{eventId}": {
      "get": {
        "tags": ["Absences"],
        "summary": "Récupérer les absences d'un événement",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des absences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {"type": "string"},
                      "comedian": {"type": "object"},
                      "organizer": {"type": "object"},
                      "reason": {"type": "string"},
                      "markedAt": {"type": "string", "format": "date-time"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}